Adjust optimization around task switching IRQ
For some reason GCC 10.3 seems to be generating code that bombs out with
`-Os` for the FIFO interrupt handler, but which works under all other
optimization settings.  Under testing it was found even the untouched
IRQ handler code from upstream had the same issue.

Force a known good optimization level around the code for now to avoid
the issue.  At some point should figure out if this is a GCC problem,
an SDK problem, or an issue in the RP2040 port.