Add portSET_IMPURE_POINTER, SMP newlib dyn-reent
To properly implement FreeRTOS SMP with Newlib while on a multicore
processor like the RP2040, multiple impure_pointers are required
(one per core) with their values set to the stored impure_ptr for
the thread currently running on that core.  This is configured in
Newlib with -D__DYNAMIC_REENT__.  It installs a weak function,
__getreent, which can be overridden to allow multiple impure_ptrs
to exist and run in parallel.

Existing code always sets _impure_ptr which cause multiple cores
to potentially update the single global _impure_ptr with bad
results.

This patch adds a macro, portSET_IMPURE_PTR, which allows a user to
define a callback to set the proper impure pointer for the current
core.

If this new macro is undefined (i.e. in a single core processor),
it will default to the simple "_impure_ptr = x" assignment.