Move SYSTICK handler to RAM, allow to halt preempt
When doing flash operations we need to block all code from flash.

Make sure the SYSTICK handler, which will be called even if all IRQs
are disabled, is in RAM.

Also add a global "don't task switch" flag, __holdPendSV, to be used
by the core to indicate it's preparing for an idleOtherCore.